Create CheckerBoard2.java
Alternate CheckerBoard class.  I made a function that would find the possible moves for a regular checker (not king), but I need to do a little more testing.  I forgot the x and y started at the bottom left in StdDraw, and need to make sure that I didn't get my coordinates confused.

For this one the key for the checkers would be their Coordinates